requirements.txt
django == 5.1.7
django-environ == 0.12.0
django-cors-headers == 4.7.0
djangorestframework == 3.15.2
pillow == 11.1.0
django-storages == 1.14.5
django-ckeditor-5 == 0.2.17
psycopg2 == 2.9.10